requirements.txt
Flask>=2.0.0
tensorflow>=2.16.0
Pillow>=8.0.0
numpy>=1.21.0
gunicorn>=20.0.0
